Add mobile mode.
Mobile keyboards emphatically do not like the document-level keyboard
handlers used by the flat mode. So, the mobile mode uses an input
field that place well with the keyboard. The CSS for the mobile view
is also more compact than the standard version.

Add filters for profanity and proper nouns.